deprecate cohp module in favour of coxx
