+++ src/lobsterpy/coxx/__init__.py
@@ -1,0 +1,7 @@
+# Copyright (c) lobsterpy development team
+# Distributed under the terms of a BSD 3-Clause "New" or "Revised" License
+
+"""This package provides the modules for analyzing COOPs/COHPs and COBIs."""
+
+from .analyze import Analysis
+from .describe import Description
